Thread the intx-src condition through every standalone launcher
The dev loop builds no dist, so a bun process needs the intx-src
condition to resolve @intx/* to source. Without it the compiled-dist
default is selected and fails. Make and the dev orchestrator already
pass it, but the launchers a developer runs by hand did not.

Wrappers that lacked the flag gain it; the entrypoints run as a bare
command gain a wrapper that supplies it. Each extensionless wrapper then
shadows its sibling source module for a relative import, so the two
in-repo importers of a wrapped entrypoint move the code they need into a
shared library and import it from there.
