+++ bin/audit
@@ -3,4 +3,4 @@ set -euo pipefail
 
 REPODIR=$(git rev-parse --show-toplevel)
 
-exec bun run "$REPODIR/bin/audit.ts" "$@"
+exec bun run --conditions=intx-src "$REPODIR/bin/audit.ts" "$@"
